+++ Nemesis-Mod/src/hardware/bluetooth/BluetoothManager.h
@@ -2,9 +2,10 @@
 #define BLUETOOTH_MANAGER_H
 
 #include <bluefruit.h>
-#include "../Packet.h"
-
-extern uint8_t UUID128_SVC_NERF_BLASTER[16];
+#include "CustomUuid.h"
+#include "CustomCharacteristics.h"
+#include "CustomServices.h"
+#include "../../Packet.h"
 
 // Provides an adapter to the onboard bluetooth module.
 class BluetoothManager {
