+++ Nemesis-Mod/src/hardware/bluetooth/CustomCharacteristics.h
@@ -1,0 +1,4 @@
+#ifndef CUSTOM_CHARACTERISTICS_H
+#define CUSTOM_CHARACTERISTICS_H
+
+#endif
