+++ Nemesis-Mod/src/hardware/bluetooth/CustomServices.h
@@ -1,0 +1,4 @@
+#ifndef CUSTOM_SERVICES_H
+#define CUSTOM_SERVICES_H
+
+#endif
