+++ Nemesis-Mod/src/hardware/bluetooth/CustomUuid.cpp
@@ -1,0 +1,3 @@
+#include "CustomUuid.h"
+
+uint8_t UUID128_SVC_NERF_BLASTER[16] = { 0x44, 0x0b, 0xf6, 0x40, 0xe8, 0xd1, 0x83, 0x49, 0xa6, 0x72, 0x4d, 0xd7, 0x72, 0xc8, 0x7c, 0x82 };
