+++ Nemesis-Mod/src/hardware/bluetooth/CustomUuid.h
@@ -1,0 +1,8 @@
+#ifndef CUSTOM_UUID_H
+#define CUSTOM_UUID_H
+
+#include <stdint.h>
+
+extern uint8_t UUID128_SVC_NERF_BLASTER[16];
+
+#endif
